Fix failing test by removing circular dependency on scigatewayRoutes
